Add support for update and initial start command
This adds support for updating an existing container, applying all
pending updates.

Included is also the initial version of the start command which will
generate all the needed configurations and start the simulation. At this
time it's not functional though all the code is there (mostly
copy/pasted from the internal NorthSec scripts).

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>